Desperate attempt to figure out how rpm should be built in 2024 to get rid of
    $ make spec
    python3 setup.py bdist_rpm --spec-only --install-script=bdist_rpm/install_script
    running bdist_rpm
    /usr/lib/python3.12/site-packages/setuptools/_distutils/dist.py:988: SetuptoolsDeprecationWarning: Deprecated command
    !!

            ********************************************************************************
            bdist_rpm is deprecated and will be removed in a future version.
            Use bdist_wheel (wheel packages) instead.

            This deprecation is overdue, please update your project and remove deprecated
            calls to avoid build errors in the future.

            See https://github.com/pypa/setuptools/issues/1988 for details.
            ********************************************************************************

    !!

Given https://github.com/pypa/setuptools/issues/1988 deprecates bdist_rpm
but does not provide any solid guidance about replacements, we just use
a .spec file template in the end. But the necessary dependencies do not
seem to be present / work on CentOS Stream 8 and 9, so we have to stop
testing on those.

We wanted to use pyproject.toml for everything.
But tool.setuptools.data-files does not respect absolute path
(unlike setup.py's data_files) so that has to be done with installs
in the .spec; and py-modules fails with
    configuration error: `tool.setuptools.py-modules[0]` must be python-module-name
    GIVEN VALUE:
        "dnf-plugins.swidtags"

    OFFENDING RULE: 'format'

    DEFINITION:
        {
            "type": "string",
            "format": "python-module-name"
        }
so we have to put those in setup.cfg.

https://gregoryszorc.com/blog/2023/10/30/my-user-experience-porting-off-setup.py/